Fix a bug with the library when displayed on screen
The library displayed an empty book when the first was eliminated. A variable was responsible for the event in deleteButton function. The variable should contain the right number of the array index to be cleared, but it always contained the index of the last element. Substitute the variable passed as a parameter in another function and the bug is solved