Fix semantic error in deleteSelf function
The function deleteSelf was created to delete the element clicked with the respective button. Though there was a problem: whenever an element was deleted the displayBook function printed a book twice. This because the function that deleted the DOM element was called after the display function. Putting the call to function later after the eventlistener solves the issue